Move EpisodicBatcher to utils.
