+++ learn2learn/data/__init__.py
@@ -7,4 +7,3 @@ A set of utilities for data & tasks loading, preprocessing, and sampling.
 from . import transforms
 from .meta_dataset import MetaDataset, UnionMetaDataset, FilteredMetaDataset
 from .task_dataset import TaskDataset, DataDescription
-from .utils import EpisodicBatcher
