+++ learn2learn/data/utils.py
@@ -44,69 +44,3 @@ def save_response_content(response, destination):
         for chunk in response.iter_content(CHUNK_SIZE):
             if chunk:  # filter out keep-alive new chunks
                 f.write(chunk)
-
-
-# Define EpisodicBatcher for Lightning algorithms
-try:
-    import pytorch_lightning as pl
-
-    class EpisodicBatcher(pl.LightningDataModule):
-
-        """docstring for __init__"""
-
-        def __init__(
-            self,
-            train_tasks,
-            validation_tasks=None,
-            test_tasks=None,
-            epoch_length=1,
-        ):
-            super(EpisodicBatcher, self).__init__()
-            self.train_tasks = train_tasks
-            if validation_tasks is None:
-                validation_tasks = train_tasks
-            self.validation_tasks = validation_tasks
-            if test_tasks is None:
-                test_tasks = validation_tasks
-            self.test_tasks = test_tasks
-            self.epoch_length = epoch_length
-
-        @staticmethod
-        def epochify(taskset, epoch_length):
-            class Epochifier(object):
-                def __init__(self, tasks, length):
-                    self.tasks = tasks
-                    self.length = length
-
-                def __getitem__(self, *args, **kwargs):
-                    return self.tasks.sample()
-
-                def __len__(self):
-                    return self.length
-
-            return Epochifier(taskset, epoch_length)
-
-        def train_dataloader(self):
-            return EpisodicBatcher.epochify(
-                self.train_tasks,
-                self.epoch_length,
-            )
-
-        def val_dataloader(self):
-            return EpisodicBatcher.epochify(
-                self.validation_tasks,
-                self.epoch_length,
-            )
-
-        def test_dataloader(self):
-            return EpisodicBatcher.epochify(
-                self.test_tasks,
-                self.epoch_length,
-            )
-
-except ImportError:
-    from learn2learn.utils import _ImportRaiser
-    EpisodicBatcher = _ImportRaiser(
-        'pytorch_lightning',
-        'pip install pytorch-lightning',
-    )
