+++ tests/unit/algorithms/lightning_anil_test.py
@@ -5,7 +5,7 @@ import torch
 import learn2learn as l2l
 import pytorch_lightning as pl
 
-from learn2learn.data import EpisodicBatcher
+from learn2learn.utils.lightning import EpisodicBatcher
 from learn2learn.algorithms import LightningANIL
